static hashmap vm3
